#227 moved native modules to packages
